WL#9635
In 7.6.7 a bug was found in the scan filtering for LCP scans that
had to be partially disabled. Now this code has been reactivated
again, it turned out that the LCP scan didn't properly check
that we didn't set the state into an illegal state. Now each
update of the page information is checked to ensure no mistakes
happen in this part of the code.