Bug #28351540: BUNDLED PROTOBUF IS OUTDATED
Description:
------------
MySQL Server source includes protobuf source code in 'extra/protobuf'
folder. The bundle is set to version 2.6.1, still it doesn't build 
properly with ubsan and other sanitizers.
The sanitizer issues are fixed in pb 3.0+, thus as workaround server
is build with system protobuf. To fix the issue, the bundle should be
updated to one of recent versions.

Fix:
----
The bundle updated to version 3.6.1. 

- adding extra/protobuf/protobuf-3.6.1
- removing extra/protobuf/protobuf-2.6.1
- adjusting configuration for new version of protobuf
- remove directories and files
   protobuf-3.6.1/benchmarks/
   protobuf-3.6.1/conformance/
   protobuf-3.6.1/editors/
   protobuf-3.6.1/examples/
   protobuf-3.6.1/objectivec/
   protobuf-3.6.1/third_party/
   protobuf-3.6.1/python/
   protobuf-3.6.1/util/
   protobuf-3.6.1/ar-lib
   protobuf-3.6.1/test-driver
- disable configuration check of cmake's mimimum version
- disable configuration zlib check
- disable installation script
- fix minor compilation issues on SunPro 5.14.0
- fix minor issues to satisfy UBSAN

Change-Id: Ie5532bacc0fb62d1ba4d524b31cbfd5fc5bb5e7d