Bug#28424366 : ADAPT CLUSTERJ TO USE CONNECTOR/J 8.0.12 FOR TESTING
Few Date, Time and Timestamp ClusterJ testcases fail, when using the
latest Connector/J. The reason for this is the complete removal of
the legacy date and time code, in the latest version of the
Connector/J. The legacy code, which was previously present and
enabled by default, prevented any conversion of the date/time types
between the mysqld and the ClusterJ test app timezones. After its
removal, the date and time types are now converted back and forth
between the mysqld and ClusterJ test app timezones, if they are
different. This becomes an issue with MTR as the mysqld is always run
in GMT and the ClusterJ test app is always run in local system time.
This conversion causes the tests to fail when the same value read
through JDBC and ClusterJ are compared and found not equal.

This patch fixes the issue by setting the session timezone of the
mysqld to the same as the timezone of the ClusterJ test app. Due to
both of them now running in same timezone, no conversion happens and
the values are intact. This patch also enables all the tests
previously disabled due to this issue. Also, the testcase
Timestamp2AsSqlTimestampTypesTest is cleaned up in this patch.

Change-Id: I4a87ac77dc29a348973f532064df74287e26150c