Bug #28526309 ER_TABLE_WITHOUT_PK CAN BE IMPROVED FOR OTHER CASES
Problem:
Message ER_TABLE_WITHOUT_PK is the same for creation of non-primary key
tables AND for dropping of primary key from a primary key enabled table,
but it doesn't give clear information that user is wrongly trying to drop the
primary key or if the action is on a non-primary-key table.

Solution:
The error message was changed to fit all three cases instead of introducing
three different error messges, that would only used once. Also cleaned up the
error message to fit the new requirements.