WL#9635
The implementation of verify checks had a problem that bits
was updated while the page was LCP scanned. To avoid this
a new bit was used to avoid updating these bits during LCP
scan of the page, instead a master bit is set and checked
at end of LCP scanned.