Fix the AutoTest test case 'testFK *'
Partly revert of *test case* changes in the previous commit:
(Bug fix in that commit is still correct)

commit e3f95e844bbe20af9a1aff5155b38c937e7a99cd
Author: Ole John Aske <ole.john.aske@oracle.com>
Date:   Mon May 14 14:54:19 2018 +0200

    Bug#27930382: FOREIGN KEY CONSTRAINT VIOLATIONS INCORRECTLY REPORTED

As part of fixing the above bug, we also intended to harden the
test case by creating a greater variety of which indexes
could be used to set up a FK-constraint. That introduced a
bug where indexes not reflecting related (self-) column were
used in the FK-constraint - Resulting in test case failing
due to key constraint violations.