Bug#28530928 ERROR 631: 'SCAN TAKE OVER ERROR', WHERE IT SHOULD NOT HAPPEN.
Considder the following case in the ACC block (access/lock manager)

* Transaction T1, have some operations in run queue.

* Transaction T2, have some operations piling up one by one in
  serial queue. Some of them need exclusive lock and blocks on T1

* All operations in run queue not part of T2 eventually completes
  or aborts. Leaving run queue empty of with only T2 operations.

* One by one, operations from T2 are put into parallel/run queue
  by Dbacc::startNext().

Dbacc::startNext() may move an OperationRec from the serial to the
parallel queue if it is allowed to run in parallel with the
other op's in this queue.

When doing so, it failed to update the Operationrec::OP_ACC_LOCK_MODE flag
which is required to reflect the accumulated OP_LOCK_MODE of all
previous operations in the parallel queue. Note that when
an OperationRec is inserted directly into the parallel queue, without
first waiting in the serial queue, this is handled by
Dbacc::placeReadInLockQueue(). In this case the OP_ACC_LOCK_MODE
flag *is* set correctly.

In turn the above inconsistency in the ACC lock queues caused
the 'scan lock take over' mechanism to fail, as it incorrectly
concluded that a lock to 'take over' was not held. The same
failure caused an assert to be hit when we aborted an operation
being a member in such an inconsistent parallel lock queue.

Patch also do some small refactoring in ::startNext() to avoid
checking 'opbits' flags where its state should already be known
due to previous checks.