Bug#28557056 ADOPT WL#12131 TO CLUSTER TESTS
As per WL#12131, use of user variable in expression is deprecated. So,
because of this some cluster test cases got affected and deprecated warnings
started appearing in their output. Hence, there is a content mismatch with the
result file and test cases failed.

Have fixed the issue by recording the deprecation warning to the test case
result file and few place where it is simple, have fixed by rewriting the
query without using user variables.