Bug #28176910	PERFORMANCE_SCHEMA.DATA_LOCKS.OBJECT_INSTANCE_BEGIN CHANGES DEPENDING ON ORDER
The main problem fixed by this patch is that engine_lock_id column of
performance_schema.data_locks was not unique, yet it was treated as such
in our code, documentation and reported CREATE TABLE.

The old format of engine_lock_id was:
trx_id:table_id for LOCK_TABLE
trx_id:space_id:page_no:heap_no for LOCK_REC

which happens to be not really unique, as a single trx can hold multiple
locks on the same table, or same row, as long as they are in different modes,
say one is LOCK_S and the other is LOCK_X.

This patch fixes this by using a new format for engine_lock_id:

trx_immutable_id:table_id:lock_immutable_id for LOCK_TABLE
trx_immutable_id:space_id:page_no:heap_no:lock_immutable_id for LOCK_REC

The newly introduced lock_immutable_id and trx_immutable_id are uint64_t
which are guaranteed not to change over a lifetime of trx and lock
respectively and also to be unique among other such objects currently in the
system. This is in contrast to trx->id which changes over time (when transaction
switches from read-only to read-write the trx->id changes from 0 to non-zero).
The reason engine_lock_id contains not only lock_immutable_id but also the
prefix, is that we need not only to identify a lock, but also be able to find
it in memory in a safe way, and this is easy to accomplish if we can first
locate the transaction which owns that lock (URI vs. URL).

This patch affects not only performance_schema.data_locks, but also other
tables which store engine_lock_id:
- performance_schema.data_lock_wait.requesting_engine_lock_id
- performance_schema.data_lock_wait.blocking_engine_lock_id
- information_schema.innodb_trx.trx_requested_lock_id

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>