Bug #27897053: ILLEGAL MIX OF COLLATIONS WHEN DATE_FORMAT                ON THE RIGHT SIDE OF OPERATION
Issue:
------
During string comparision, we coerce the right argument to
the left's collation.

Solution:
---------
Use the appropriate aggregations to try to find a superset.

Also, make sure UTF32 is counted as a superset of any
collation. Previously, only utf8mb4 would have this status.

This is a backport of Bug#25642319 and Bug#27452148.