Bug#28575934 NDB_MY_ERROR() SEGV FOR MAPPED ERRORS
The function ndb_my_error() is used in
ndbcluster_get_tablespace_statistics() and prepare_inplace_alter_table()
to report errors when the function fails to interact with NDB. The
function is expected to push the NDB error as warning on the stack and
then set an error by translating the NDB error to a MySQL error and then
finally call my_error() with the translated error. When calling
my_error(), the function will extract a format string that may
contain placeholders and use the format string in a sprintf like
function. Since my_error() is called without arguments the sprintf like
function may read arbitrary memory and a segmentation fault occurs.

Fix by changing tactics to always push the NDB error as warning and
then set an error with a provided message. Add new helper function
to Thd_ndb to be used in favor of ndb_my_error().

Change-Id: I30c37d0f683b11b9279496c26f5b0855592e00d5