BUG#28564032-AUTOTEST REPORTS TESTRUNS WRONGLY
Autotest reports number of test runs in case of failure testcases correctly.

Existing logic works with a loop where the testruns will be increased in a
loop in order to acheive the maximum number of test runs if user wish to retry in
case of test case failure. This causes the testruns to be incremented to have a
conditional check, hence causing the wrong entry of testruns in report file.

Changes made to reduce the testruns by one in case of test failure in order to get
correct entry of testruns in report file.