Bug#27820277 - MYSQL SERVER CRASH WHILE SET PASSWORD
DESCRIPTION
===========
When a user is deleted from mysql.user which is followed by
creating few users and finally SET PASSWORD is fired, it
results in server exit.

ANALYSIS
========
Each time we execute CREATE USER .., it pushes the user to
in-memory acl_users list. When the size of the list equals
the capacity, it mallocs a new buffer with twice the
capacity, copies all the elements and frees the old buffer.
This freed stale memory is returned by find_acl_user()
during SET PASSWORD and later accessed in
update_sctx_cache() resulting in the issue.

replace_user_table() calls ha_index_read_idx_map(), which
tells if or not a given user exists in mysql.user table.
When SET PASSWORD is called, this func throws an error which
signifies that the user is not in the mysql table (as it
was deleted in the very 1st step)

As of now we're not handling this error and proceeding. We
can avoid the reported issue altogether if we handle the
error and return from that point.

FIX
===
Above condition is handled in replace_user_table(). Also
that the call to this function is modified in
change_password() and mysql_alter_user(). The parameter
'can_create_user' is passed as false instead of true so
that its handled accordingly in the patch.