Bug#28585914 ROUTER TEST ISSUES REPORTED FOR PB2 8.0.13 RELEASE BRANCH - SOLARIS X86
There seems to be a compiler issue that is making the binary SEGFAULT
each time std::condition_variable::wait_for gets called.
This can be observed on a simple example:

std::condition_variable cv;
std::mutex cv_m;
int main()
{
  std::unique_lock<std::mutex> lk(cv_m);
  cv.wait_for(lk, std::chrono::milliseconds(10), [](){return true;});
}

If you compile it with:
$ /opt/developerstudio12.6/bin/CC -std=c++11 -m64 -xO2 test.cc -o testapp
you'll get a crash when running the binary.

-xO1 doesn't crash
-xO2 crashes
-xO3 doesn't crash
-xO4 doesn't crash
-xO5 doesn't crash

This patch is really only a workaround to change the optimization level to -xO1
for the router sources to avoid this issue.

Reviewed-by: Tor Didriksen <tor.didriksen@oracle.com>
RB: #20426