Bug #28595557: USE-AFTER-FREE IN TEMPTABLE ALLOCATOR
Given certain patterns of allocations, copies/rebinds of the allocator and
deallocations, temptable::Allocator can reuse a freed memory block. In
particular, this leads to crashes in the test suite when Clang on Windows
uses the Visual Studio 2017 STL and tries to run anything including
common-tests.inc. (It possibly also affects Visual Studio 2017 with the MSVC
compiler, but I haven't verified.)

Fix by putting the state into a shared_ptr, shared between all instances of
a given allocator; this is arguably a kludge, but it's hard to do better given
the current design. This also fixes the issue where the number of allocated
blocks could go negative.

Change-Id: I1dabdbed533e199099770195ebe56b8fccb7f34f