BUG#27966483 - SIGHUP CAUSE MYSQL SERVER CRASH.
When SIGHUP signal is sent to mysqld and the server
is admitting new connections at this time, it results
in the server crash. The thread cache is purged as part
of SIGHUP handling. This involves raising signal to the
threads that are blocked waiting for new connections to be
in thread cache. We purge the channel queue as part of this
making it empty. New connections can be admitted after which
result in waking up a thread in the thread cache and this attempts
to dequeue a connection for handling from an empty channel info queue.
This result in the behaviour bug.
The fix is not to drain the channel info queue in the kill_blocked_pthreads
and instead let it be handled by the unblocked thread.