Bug#28589637: ED_CONNECTION CANNOT EXECUTE QUERIES THAT RETURNS A RESULTSET
For each result set recieved, Protocol_local stores the metadata
and the actual results seperately. At the end of every result set, a new
instance of Ed_result_set is created with the metadata and the actual
results. This instance is then added to a list of result sets stored in
Ed_connection.

Methods are provided in Ed_connection to retrieve the resultsets and
fields.

Change-Id: I62abbdb6631b726611bb8c9ef7f3aa83adad5fe1