WL #12470: Volcano iterator semijoin [3/5, duplicate weedout]
Weedout is generally rewritten into a new iterator called WeedoutIterator
that reuses most of the logic from do_sj_dups_weedout(), and which runs
after the joins.

There is one complicating factor in that the join optimizer can sometimes
set up tables such that two weedout operations overlap, and then combine
the two into one larger weedout. In this case, we can get extra unlucky
and get into a non-hierarchical situation where a weedout operation spans
two outer joins. Since this should be exceedingly rare in practice, we
solve it by moving the weedout to the very top of the join.

Change-Id: I7394d67eecdc2d8bf88e8ecbb10b8a5b7f507f6b