BUG#28679449 IMPROVES PROCESS'S PID FILE DELETION
Deletes pid file when a temporary process is in state "running" but is
no longer running in the system.
Do not remove pid file when scheduling the process to stop, allowing it
to be removed only when the process is actually stopped.

These changes were identified while chatting with Mauritz about on-going
investigation of autotest (atrt+ndb_cpcd) failures. One of the known
issues was pid files not being removed while the processes were long
gone. This was due to not accounting that a running process could stop
for reasons others then CPCD stopping it (e.g. server process crash or
test finishing).
The removal of the pid file when a process is scheduled to be stopped
seems unnatural, because we have explict code to remove the file when
the process is actually stopped - so code was removed.