Bug 28622708 - SERVER START FAILS TRYING TO RE-CREATE UNDO TABLESPACES
There was an error in string processing in a function called
Fil_path::is_undo_tablespace_name() when the undo file name had
a directory attached to it. This happened because
--innodb-directories=x and --innodb-undo-directory=x/undo.
So that function was trying to parse "undo/undo001" instead of just
"undo001" like it normally does.  Since it did not recognize this
as an undo file, it did not open it at file discovery.  Then later,
InnoDB would try to create the file, but it was already there.

Along with that bug testing showed that --innodb-undo-directory
could be assigned a directory above the datadir.  This would
would increase the scope of files that are scanned at startup
duriong file discovery.  A new error is added to prevent the
undo directory from being an ancestor of the datadir.
--innodb-directories can still specify ancestors of the datadir.

Approved by Rahul Agarkar in RB#20523