BUG#28637895: "EXEC" COMMAND IN MYSQLTEST DOESN'T RETRUN OUTPUT IN               CASE OF FAILURE
Description:
-----------
Mysqltest "exec" command was refactored under BUG#28358835. Before the
bug fix, when a subprocess started by "--exec" command fails, mysqltest
used to print the output of the subprocess. Having this output in case
of "exec" failure, allows user to investigate the cause for the failure.

Fix:
----
When a command executed using "exec" fails, print the output the command
on console.

Change-Id: I7904bf0b2b58b7d07e323f3957f05130c1e7bc10