Bug#28642289 MYSQL_SERVER_MOCK DOES NOT SHUTDOWN GRACEFULLY WHEN KILLED ON WINDOWS.
In the Router component tests when we try to stop the child process
we use kill() method. This method on Windows is implemented with
the call to TerminateProcess(). That does not allow the process
that is being killed to do a graceful shutdown, like closing
the open sockets etc... That has an impact on the tests stability
as for example trying to connect to the port that killed process
was using takes long time to timeout while it should fail
immediately.

This patch replaces use of TerminateProcess() with sending
Ctrl+Break singal and handling it in the child process
to emulate SIGTERM behavior.