BUG#28516776 NDBIMPORT REMOVE LOGX DEFINES
Here we are dealing with diagnostic logging messages in a form like

  log(level, message)

As a general rule here, we prefer not to evaluate the expression
"message" unless the current diagnostic logging level is set to
"level" or higher. "message" may be expensive to evaluate, and
log statements may occur in tight inner loops of the program logic.

The bug report asks to remove the macro, but using a macro is perhaps
the best way to prevent the message from being evaluated unnecessarily.

Nonetheless, we can definitely improve the macro by moving its mutex and
timer calls somewhere else. This patch moves them into special "start" and
"stop" tokens in the message.

As a side effect, the elapsed time is moved to the leftmost position in
the log message and displayed in a fixed-width format.