BUG#28516776 NDBIMPORT REMOVE LOGX DEFINES
One problem with the macro names log1(), log2() etc. is that log2
may also be the name of a math function.

In this patch, log1(msg) and log2(msg) are universally replaced with
log_debug(1, msg) and log_debug(2, msg). This prevents any conflict
with a math library, and removes one level of preprocessor indirection.

log3(msg), which is conditionally defined at compile-time, is changed
to log_debug_3(msg).

log4(msg), which is not used in the code, is removed.