WL#12374 ATRT SUPPORTS MAX-RETRIES COMMAND-LINE ARGUMENT
Adds support to ATRT to allow "--max-retries" as command line argument.

Modifies ATRT to support "--max-retries" command-line argument option
to set the default value of max-retries. If no option is set, the default value is zero.
Default value or value set in command-line argument will always be overwritten with
the value specified in the testsuite.