BUG#28616442 FIX TO BUG#28062382 DRAMATICALLY REDUCES THROUGHPUT WHEN              THE BINLOG IS ACTIVE
The fix to BUG#28062382 was saving CPU by doing longer waits in log threads,
when there was nothing to be done for the log threads and the frequency of
requests to write / flush redo was detected as very low. The longer waits
meant: no spin-delay and initial timeout = 10ms.

The log threads were waiting on their events (cond_var), so if there was
a task for them, user threads were trying to wake them up.

The problem was because user threads were trying to find out if they
should wake up log threads. This leaded to very complex logic which
has already proved to be error prone.

The whole logic became simplified in this patch. User thread checks
if write_lsn is not advanced enough and wakes up log_writer in such
case - no matter if frequency is high/low or CPU usage on server is
below 80% threshold or not.

There was also a bug in monitor of the frequency - when there were
no counted calls to log_write_up_to(), the average time between calls
stayed unchanged. It should be reset to +inf if that happened for
longer period.

Additionally all calls to log_write_up_to() should always be counted.
The calls from page cleaners were the only that we wanted to exclude
from being counted, so page cleaners should not call log_write_up_to()
if the flushed_to_disk_lsn is already advanced enough (most cases).