WL#9270 - X Protocol Prepared Statement execution
Applications that repeatedly execute the same kind of statements with changed values,
currently have to go through same parse-execute cycle for every execution.
In the case of the Document Store a large part of the queries are simple _id based lookups
which can be accelerated by preparing the Crud once and executing it often.

Approved by Vinay Fisrekar <vinay.fisrekar@oracle.com>