BUG#28663647 WHEN MAX_BINLOG_SIZE IS SET, EXTRA 512 BYTES OF THE HEADER IS NOT CONSIDERED
Problem
=======

The extra encryption header size was not being taken into account when
checking if a binary or relay log file reached the max_binlog_size (or
max_relay_log_size).

Analysis
========

The rotation by size was relying on plain binlog stream position and not
in the real file size.

Fix
===

Created a new function get_real_file_size() to Binlog_ofile and used it
instead of position() when checking the file size for automatic
rotation.