Bug#28460158 SIG 11 IN ITEM_FUNC_GET_DD_CREATE_OPTIONS::VAL_STR AT SQL/ITEM_STRFUNC.CC:4167
The problem is that the INFORMATION_SCHEMA query is being invoked
with invalid key-value pair string being parsed when building
dd::Properties object by parse_properties(). We fail to allocate
a dd::Properties object upon parse failure and return nullptr.
The problem is we are deferencing a nullptr.  This problem is hit
in rare situation during RQG run.

This patch reports a warning to server error log in such
situations and avoids dereferencing nullptr.

Change-Id: Ic1d703d1a6e22f7942b41f99f3cd25f19d078808