Bug #28284624 `PURGE BINARY LOGS TO` FAILS WHEN BINARY LOG               SPECIFIED IS NOT IN BINLOG BASE DIR
Problem
-------
`PURGE BINARY LOGS TO` fails if binary log is not present in
the data directory.

Analysis
--------
When the binary logs are being moved to another location by
mysqlbinlogmove, the index file is also updated with the
absolute paths of the binary logs.

When the server is executing `PURGE BINARY LOGS TO` command,
it ends up in comparing the relative path of the 'to_log'
against the absolute path in the index file instead of
comparing the basenames of the files.

Fix
---
Added a wrapper function to compare the log names.