Bug#28647699 SOME NDB TESTS ARE FAILING IN PB2 DUE TO LIBRARY MISMATCH
Most of the ndb api test executable uses mysql and cluster dynamic libraries.
But as the rpath is not set, during runtime the linker will look for the
libraries only in LD_LIBRARY_PATH and it throws library not found error.

Modifed the make file to add the path of shared libraries to the executable.