BUG#27463277: scalability issue with srv_session in mutex_map_thd_srv_session
In order to speed up the query throughput when using the x plugin some
changes had to be made to the way threads were handled. Prior to this
patch a Srv_session would attach and detach to and from the current
thread for every command, creating a hotspot. With this patch the thread
switch only happens if we intend to disconnect the current session, or
if the current thread and the one we are suppose to attach to belong to
different plugins.
In order to achieve this a setter and a getter for the plugin id was
added to THD and the algorithm in switching threads was changed to take
into consideration if the current plugin and the one to switch to belong
to the same plugin pool of threads.