Bug#28526737: SECONDARY ELECTION PROCESS CONTINUES WHEN IT SHOULD ABORT
After the fix of
BUG#28401703: USE THE NEW SHUTDOWN SERVICE WHEN GR_EXIT_STATE_ACTION=ABORT_SERVER
the abort process is not instantaneous.

As a result, the secondary election process fails to set the read mode, the
process invokes the abort but continues.
To the user, it appears the changes to single primary mode for example are
able to progress and even terminate locally when they shouldn't

Resolution:
Code was not handling error state assuming SIGKILL will happen.
Code has been modified to set error state while shutdown happens.

Additional NOTE: It will not be pushed to 8.0(if pushed by 24th August).
27th Onwards present trunk will be 8.0 so its already tested.