WL#12137: ENHANCE MTR TO DISABLE/ENABLE SECONDARY STORAGE ENGINE
Description:
------------
MySQL now supports secondary storage engine for a table.

This WL will introduce a new MTR option --secondary-engine

  * When --secondary-engine option is enabled, MTR will take care of
    doing the necessary setup needed to support the secondary engine.

WL#12114: MTR SUPPORT FOR AUTOMATIC LOADING OF DATA TO SECONDARY
          ENGINE AFTER A DDL STATEMENT

Description:
------------
This WL will introduce a new MTR option --change-propagation. This
option must be used along with --secondary-engine option.

When this option is defined or enabled, MTR will automatically change
the secondary engine of a table and will load the table contents from
primary engine to secondary engine after a DDL statement.

The scope of this WL is limited to following DDL statements

  1. ALTER TABLE (ALTER TABLE ... RENAME newtab not supported)
  2. CREATE TABLE
  3. CREATE INDEX
  4. DROP INDEX

WL#12113: MTR SUPPORT FOR AUTOMATIC LOADING OF DATA TO SECONDARY
          ENGINE AFTER A DML STATEMENT

Descrption:
-----------
This WL will enhance --change-propagation option to take an optional
argument (i.e either 0 or 1). Setting this option value to 1 is same
as specifying the option without an argument.

When the value is set to 0, MTR will automatically unload and then
load the data back to secondary engine after a DML statement.

The scope of this WL is limited to following DML statements

  1. DELETE
  2. INSERT
  3. REPLACE
  4. UPDATE

WL#12366: MTR: LOAD THE TABLE CONTENTS TO SECONDARY ENGINE AFTER A
          SERVER RESTART WITHIN A TEST

Description:
------------
When a mysqld server is restarted within a test, even secondary engine
server is restarted if enabled. This will cause unloading of the data
from secondary engine and hence the queries after the server restart
won't run on secondary engine.

This WL will enhance MTR to support automatic loading of tables from
primary engine to secondary engine after a server restart within a
test.

WL#12367: MTR: GENERATE A REPORT OF NUMBER OF QUERIES RUN ON SECONDARY
          ENGINE

Description:
------------
This WL will enhance MTR to generate a report containing information
about number of statements run on secondary engine server (i.e value
of Secondary_engine_execution_count status variable) for each test
when a MTR run is started with --secondary-engine option.

Change-Id: Ic632cbd9d0ed71bb2e24db189d913e5da08d43e4