Bug #28615893 CRASH OF MYSQLD WITH 128K JSON FOR JSON_REPLACE THROUGH SYSBENCH
Problem:

There is problem in the approach taken to allocate a fragment from the many
fragment pages available.  First we search for a fragment page with sufficient
space available for the given payload.  Once a fragment page is found, then we
allocate a fragment in this page.  We assert this this allocation must succeed.
But there are valid situations in which the fragment doesn't fit in the given
fragment page.

Solution:

Instead of asserting, continue to try for fragment allocation in other fragment
pages.

rb#20503 approved by Mayank.