WL#11123: Group Replication: hold reads and writes when the new primary has replication backlog to apply
This worklog implements a fencing mechanism when a new primary is being
promoted in Group Replication (GR). The fencing will restrict connections from
writing and reading from the new primary until it has applied all the pending
backlog of changes that came from the old primary. Applications will not read
stale data or do writes for a short period of time (during the new primary
promotion).