Bug#28541204 SECURE-FILE-PRIV BREAKS LOAD DATA INFILE              REPLICATION IN STATEMENT MODE ON 5.7.23
Problem
-------
SBR breaks when LOAD DATA is executed on 5.7.23 server.

Analysis
--------
As a part of a bug fix that went into 5.7.23 where MySQL-5.7
was made to compile warning-free, the if condition in
sql_load.cc was changed from

5.7.22: if (thd->slave_thread & ((SYSTEM_THREAD_SLAVE_SQL |
            (SYSTEM_THREAD_SLAVE_WORKER)) !=0))

to

5.7.23: if ((thd->slave_thread & (SYSTEM_THREAD_SLAVE_SQL |
            SYSTEM_THREAD_SLAVE_WORKER)) != 0).

While the former condition evaluated to true, after the
change, the latter always evaluated to false thereby causing
the slave to fail.

Fix
---
Replaced thd->slave_thread with thd->system_thread enum.