BUG#28616442 FIX TO BUG#28062382 DRAMATICALLY REDUCES THROUGHPUT WHEN              THE BINLOG IS ACTIVE
Eventum 61197 revealed performance regression introduced for DDL intensive
workloads when innodb_flush_log_at_trx_commit = 2.

This is because we over-simplified in the previous fix for BUG#28616442.

This patch reverts related part of the previous fix:
- don't sleep for more than 1ms in log threads (trx=0 case could suffer),
- when trx=2 and user thread is waiting for flushed redo (ddl), we need
  to first wait for redo written and wake up log_flusher only afterwards
  (otherwise we could wake up log_flusher too early - that was problem).

We also skip disabling spin-delays for two cases:
- waiting for written redo when trx=1,
- waiting for flushed redo when trx=2.

We also checked if we could use two-steps waiting always - that is always
first wait for written redo and only afterwards fallback to wait on
flushed redo. We can't afford that.

This would result in more CS when busy-waiting is disabled and trx=1
(it's disabled when total cpu usage is more than the hwm (default 50%)).

However we observed that two-steps waiting increased TPS for workloads
with small number of connections. We confirmed that it was because of
increased total time for busy-waiting (two times more spin rounds).
On the other hand, we can't afford making more spin rounds for higher
number of connections, because we would risk wasting too much cpu,
before we reach the hwm which disables the busy waiting.

Because of that we introduced new mechanism:
1. Use increased spin rounds (compute in range 250k..25k) when
   cpu usage is below 50% of the hwm.
2. Use default spin rounds (25k) when cpu usage is in 50%..100% of hwm.
3. Do not use busy waiting when cpu usage is higher than hwm.

Points 2,3 - we behave as it was in the past.

Point 1 - we may use increased number of spin rounds when cpu usage
is lower than 50% of the hwm.