Bug #26925260 ERROR 1030 (HY000): GOT ERROR 44 - 'INNODB ERROR' FROM STORAGE ENGINE
PROBLEM

In linux , partition suffixes eg "#p" are not converted to lowercase
when lowercase_table_names is set to 1 . So when some tablespace files
are moved to linux system from windows, an OS rename call can fail
thinking the source file to be renamed doesn't exist.

FIX

if lowercase_table_names is set to 1, convert the part of the tablename
containing  partition suffixes to lowercase as well.

Reviewed by: Bin Su <bin.x.su@oracle.com>
	     Aditya A <aditya.a@oracle.com>
RB: 20591