WL#461 allow non-LATERAL outer references in derived table (SQL92?) example: select * from t1 where t1.a = (select dt.b from (select t3.c*t1.d from t3) as dt(b));                                             ^^^^ the reference to t1.d is now allowed. If "dt" is materialized, it will be re-materialized for every execution of the query block which contains the definition of "dt" (here, that means the =(...) subquery). "dt" may also be merged. Also possible in a CTEs: select * from t1 where t1.a = (with dt(b) as (select t3.c*t1.d from t3) select dt.b from dt);                                       ^^^^ If the CTE is materialized, it will be re-materialized for every execution of the query block which contains the definition of "dt" (here, that means the =(...) subquery). The word "definition" matters as a CTE is defined in a place and possibly used elsewhere.
WL#8652 allow LATERAL derived tables (SQL:99)
example:
select * from t1,
lateral (select t3.c*t1.d from t3) as dt(b));
                     ^^^^
the reference to t1.d is now allowed. If "dt" is materialized, it
will be re-materialized for every row read from t1.
"dt" may also be merged.
A CTE must not have the LATERAL keyword.

The difference between the two WLs is the origin of t1.d:
a non-lateral derived table DT can reference tables which are
already read when the subquery containing the definition of DT starts
executing;
a lateral derived table DT can also reference tables which are not
already read when the subquery containing the definition of DT starts
executing: tables which are read only right before when the nested-loop-join
execution comes to the stage of reading DT.
Rematerialization of lateral table is mentioned in EXPLAIN and opt trace,
with a mention of when in the nested-loop this happens.

Non-obvious changes:
- mark_select_range_as_dependent() was unused, removed.
- changed resolution code in Item_field::fix_outer_field() and Item_ref::fix_fields()
to allow derived tables to find outer references: the places where they
can search are a _strict_ subset of "all subqueries above".
- a SELECT_LEX_UNIT underlying a derived table keeps track of its
lateral outer refs (with table bits); if it has any outer ref it uses
flag UNCACHEABLE_DEPENDENT.
- order of merging of derived tables is changed; used to be:
merge all possible - materialize rest; now it's:
merge one if possible - if not materialize it; go to next table.
- when we merge a non-lateral derived table having an outer ref,
it may make it lateral, if the outer ref is now on same FROM clause.
- place of lateral table in query plan is decided cost-based based
on the table is depends on (to minimize the number of rematerializations);
this implies keeping track of dependencies-already-in-join-prefix
when in greedy search.
- 'got_final_plan' introduced to simplify functions.
- in planning, join buffering is disabled for all tables between a lateral
table (incl.) and the first of its dependencies (excl.), so that rows
of that first dependency don't get shuffled (to minimize the number of
rematerializations).
- printing of ON conditions in opt trace / EXPLAIN was sometimes missing
(EXPLAIN even showed a one-table query when it was actually a left join)
- TABLE::empty_result_table() contains commonly used code.