Bug#28684772: IS.INNODB_TABLESPACES NOT SHOWING TABLESPACE INFO AFTER UPGRADE(8.0.12->8.0.14) Bug#28684204: UPGRADE FROM 8.0.12 TO 8.0.14 FAILS WHEN --INNODB_UNDO_TABLESPACES > 2
There were two problems when upgrading from 8.0.12 to 8.0.14.
1) dd::Tablespaces::se_private_date is missing key/value pairs for
the key 'state' and it may contain a kv pair for 'discarded.
2) The WL9508 code assumed that there can only be 2 implicit undo
tablespaces. But 8.0.12 allowed and used the setting
--innodb-undo-tablespaces to create up to 127 undo tablespaces.

The solution is to make the "state" key optional and add back in the
possibility of key "discarded".
Make all calls to get the current tablespace state go through the same
function, which uses either dd::tablespaces::se_private_data or the
'discarded' key, or the undo::Tablespaces object in memory.
Treat implicit undo tablespaces with undo number > 2 as an explicit
undo space in that they can be made inactive and empty (as the code
worked before, but they also can be dropped.
There were other places in code where these guidelines needed to be
enforced as well.

Approved by Rahul Agarker in RB#20647.