Bug #27453180 FOREIGN KEYS CONSTRAINTS IGNORED AFTER RENAME TABLE
Problem:

After a RENAME TABLE operation, some of the already defined foreign key
constraints were not taken into account.  This is because the foreign
keys are only associated to the new name given to the table and are not
related to the old name of the table.

Solution:

During RENAME TABLE operation load the foreign key constraints associated
with the new table name unconditionally.

rb#20658 approved by Mayank