Bug#28554157: TABLES UPGRADED FROM 8.0.11 ALGORITHM=INSTANT DOESN'T WORK UNLESS YOU REBUILD IT
Problem: ALGORITHM=instant was rejected incorrectly on tables created in
a version prior to 8.0.12. Root cause turned out to be that beginning in
8.0.12 some key_info flags were no longer default when creating a table. This
meant that when the inplace/instant analysis compared the old key
definitions to that of the intermediate table created during alter, it
incorrectly concluded that all indexes had changed. This then lead to
the request for algo=instant being rejected.

Solution: Ideally this kind of thing should be fixed during upgrade, but
since we don't have the infrasructure for that yet, we choose to omit
these flags when loading the TABLE_SHARE from the DD. This fixes the
alter problems, but the options column in the indexes table will contain
incorrect information until the table is rebuilt.

Change-Id: I5b42f00c10a0682f3cfc7c66146eab7891a58266