Bug#28714864: Running 'analyze' on table with unsupported index length crashes the data nodes
Maximum length of an index supported by mysql is 3072. So, in the
ndb_index_stat_sample table, as it is visible also in the mysql server,
the primary key size cannot exceed 3072 bytes. And the primary key
consists of 4 attributes - index_id, index_version, sample_version,
stat_key. The first three are just 4 bytes(1 word) each. And one more
word for the stat_key var bytes length. That leaves us (3072 - 4 * 4 =
)3056 bytes for storing the actual stat_key data. Due to this the
maximum supported length of an index by the ndb index stat is 3056
bytes. But when a table with a key having more than that is created, as
mysqld allows up until 3072 bytes for an index, we get undesirable
results. The data nodes, are equipped to handle only key size of upto
3056 in the ndb_index_stat_sample table. So when the data node tries to
generate statistics on an index whose max key length is more than 3056,
it fails to handle them properly and crashes.

This patch fixes that by checking the index length in the very beginning
of the statistics generation process and returning an error if the index
has an out of range key length. The error was not properly forwarded to
the client due to some missing parts in the signal flow. This patch
fixes them also.

Changes:
 AccScan.hpp
  - Fix the misaligned old error codes
  - Add a new error code for invalid index key length
 DbtuxScan.cpp
  - Add code to check if the index has a supported key length and return
    error if required
 Suma.cpp
  - Fix the GSN_SCAN_FRAGREF signal execution and forward the error code
    returned
 Trix.cpp
  - Forward the error code properly to the client during SUB_SYNC_REF
    execution
  - Fix the statCleanExecute and statScanExecute to calculate right maximum
    of the after values
 ndberror.c
  - Add the new error and a missing old error definitions to file
 NdbIndexStat.hpp
  - Add the new error code
 sql/ha_ndbcluster.cc
  - Fix the ha_ndbcluster::analyze function to properly handle the
    error returned to it by the statistics functions
 sql/ha_ndb_index_stat.cc
  - Handle the new error in ndb_index_stat thread
 ndb_index_stat.test
  - Add few testcases to test how the data nodes handle index with large
    key lengths