Bug#27613173 - UNDEFINED BEHAVIOUR IN SQLD WHEN THE NDB_INDEX_STAT_SAMPLE VALUES ARE CORRUPT
The ndb_index_stat thread reads the ndb_index_stat_sample table when
it detects a change in the index values. If for some reason the
ndb_index_stat_sample table values are corrupt, the read by the
ndb_index_stat thread ends up corrupting random memory locations in the
mysqld.

This is because the ndb_index_stat thread always provides a buffer with
a pre-calculated, pre-allocated size based on the number of attributes
in the respective index rather than providing a buffer with the length
of the ndb_index_stat_sample's maximum attribute lengths for receiving
the data. The values of the ndb_index_stat_sample calculated by the
kernel are usually of the exact same size precalculated by the mysqld.
But when the ndb_index_stat_sample table has corrupted values, this is
not true. And when the corrupted value received by the mysqld is of
greater length than the buffer provided, memory gets overwritten and
random crashes happen. This patch fixes that by always providing a
buffer of maximum length for copy out. Also, the length of the data
received is checked and if in case the length is greater than what was
expected, an error is reported and the key and value length are printed
to the error log. This should help us further investigate why there are
bad values in the first place.

Change-Id: Id45ddad957a9055901b61b2f1c801213eb75c3e6