Bug #28380808 NDB : GRACEFUL STOP API STATE NOTIFICATION TOO SLOW
APIs should observe that a node is moving through SL_STOPPING
phases and stop using the node for new transactions.  This
minimises the disruption in the later phases of the node stopping.

However, APIs are only informed of node state changes via periodic
heartbeat signals, so may not have the chance to take avoiding
action.

Where the heartbeat interval is high, APIs continue to attempt
to use nodes which are shutting down, generating unnecessary
failures.

Solution is to directly notify all APIs when a data node is
being gracefully stopped.  This allows them to experience
minimal disruption.